% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{AddColor}
\alias{AddColor}
\title{Add colors and styles for printing a vector of objects.
The function returns a vector of character strings,
cyan-colored normal-style strings by default.}
\usage{
AddColor(x, color = "cyan", style = "default")
}
\description{
Add colors and styles for printing a vector of objects.
The function returns a vector of character strings,
cyan-colored normal-style strings by default.
}
