% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{CheckSep}
\alias{CheckSep}
\title{In an annotation file, the value in each column are often
delimited by symbols like ';', ',', or '|'. This function
is for finding the proper delimiter in the given strings.
TODO: The problem on whether the reg expr is fixed or not.}
\usage{
CheckSep(x, priority = c(";", ",", "\\\\|"), default = "\\t")
}
\description{
In an annotation file, the value in each column are often
delimited by symbols like ';', ',', or '|'. This function
is for finding the proper delimiter in the given strings.
TODO: The problem on whether the reg expr is fixed or not.
}
