% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadLines}
\alias{ReadLines}
\title{ReadLines() reads the input file or text and returns a vector of lines.
The param "delimiter" is for XLSX2Lines().}
\usage{
ReadLines(file = NULL, text = NULL, delimiter = "\\t",
  func.name = "ReadLines()")
}
\description{
ReadLines() reads the input file or text and returns a vector of lines.
The param "delimiter" is for XLSX2Lines().
}
