% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadTable}
\alias{ReadTable}
\title{ReadTable() reads the input file or text and returns a data frame.}
\usage{
ReadTable(file = NULL, text = NULL, header = FALSE,
  delimiter = "\\t", check.names = FALSE, comment.char = "",
  func.name = "ReadTable()")
}
\description{
ReadTable() reads the input file or text and returns a data frame.
}
