% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadText}
\alias{ReadText}
\title{ReadText() reads the input file or text and returns a string of text.
The function wraps ReadLines() around and
add a "\n" as the EOF for the output text.
The param "delimiter" is for XLSX2Lines().}
\usage{
ReadText(file = NULL, text = NULL, delimiter = "\\t",
  func.name = "ReadText()")
}
\description{
ReadText() reads the input file or text and returns a string of text.
The function wraps ReadLines() around and
add a "\n" as the EOF for the output text.
The param "delimiter" is for XLSX2Lines().
}
