% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{WriteLog}
\alias{WriteLog}
\title{Print messages, warnings, or errors in a unified format.
The messages will be written to STDOUT.
And the warnings and the errors will be written to STDERR.}
\usage{
WriteLog(text, func.name = "Default", level = c("info", "warn",
  "error")[1], color = NULL, style = NULL, saveWhenQuit = "default")
}
\description{
Print messages, warnings, or errors in a unified format.
The messages will be written to STDOUT.
And the warnings and the errors will be written to STDERR.
}
