% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{XLSX2Lines}
\alias{XLSX2Lines}
\title{XLSX2Lines() reads the given xlsx file and returns a vector of lines.}
\usage{
XLSX2Lines(file = NULL, delimiter = "\\t")
}
\description{
XLSX2Lines() reads the given xlsx file and returns a vector of lines.
}
